sidebar UI수정
